requirements.txt
imgaug
opencc
easydict
opencv-python
pillow
numpy
glob2
#json
#random
#string